src: infrastructure: added a new GenericRepository implementation to store slices on the host file system
